pxtextmining/requirements.txt
absl-py==0.11.0
alabaster==0.7.12
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.2.1
blis==0.4.1
bokeh==2.2.3
brotlipy==0.7.0
CacheControl==0.12.6
cachetools==4.1.1
cachy==0.3.0
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
cleo==0.8.1
click==7.1.2
clikit==0.6.2
cloudpickle==1.6.0
colorama==0.4.4
crashtest==0.3.1
cryptography==3.1.1
cycler==0.10.0
cymem==2.0.3
cython==0.29.14
cytoolz==0.11.0
dask==2021.1.1
dask-glm==0.2.0
dask-ml==1.8.0
dask-xgboost==0.1.11
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
distlib==0.3.1
distributed==2021.1.1
docopt==0.6.2
docutils==0.16
emojis==0.6.0
en-core-web-lg==2.3.1
en-core-web-sm==2.3.1
entrypoints==0.3
feather-format==0.4.1
filelock==3.0.12
flake8==3.8.4
fsspec==0.8.3
future==0.18.2
gast==0.3.3
gensim==3.8.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
idna==2.10
imagesize==1.2.0
imbalanced-learn==0.7.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.17.1
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.0.2
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-tuner==1.0.1
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
llvmlite==0.36.0
local-pxtextmining==0.1.0
locket==0.2.1
lockfile==0.12.2
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.7.0
msgpack==1.0.2
multipledispatch==0.6.0
murmurhash==1.0.2
mysql-connector-python==8.0.23
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
nltk==3.5
notebook==6.1.4
numba==0.53.1
numpy==1.20.2
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandas-datareader==0.9.0
pandas-profiling==1.4.1
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.7.0
partd==1.1.0
pastel==0.2.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==21.0.1
pipreqs==0.4.10
pkginfo==1.7.0
plac==0.9.6
pluggy==0.13.1
poetry==1.1.6
poetry-core==1.0.3
preshed==3.0.2
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.7.0
pxtextmining==0.1.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
pylev==1.3.0
pylint==2.6.0
PyMySQL==1.0.2
PyNaCl==1.4.0
pynisher==0.6.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytz==2021.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.18.0
rpy2==3.3.6
rsa==4.6
Rtree==0.9.4
sacremoses==0.0.44
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
setuptools==50.3.1.post20201107
shellingham==1.4.0
sip==4.19.13
six==1.15.0
sklearn-pandas==2.1.0
smart-open==3.0.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
spacy==2.3.5
spacy-lookups-data==0.3.0
Sphinx==3.5.3
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.23
srsly==2.4.0
tabulate==0.8.7
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
textblob==0.15.3
thinc==7.4.1
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.1
tomlkit==0.7.0
toolz==0.11.1
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
transformers==4.5.0
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.1
UNKNOWN==0.0.0
urllib3==1.25.11
vaderSentiment==3.3.2
virtualenv==20.4.4
wasabi==0.8.2
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xgboost==0.90
yapf==0.30.0
yarg==0.1.9
zict==2.0.0
zipp==3.4.0
